requirements.txt
 python-dotenv==1.1.0
 mistralai==1.6.0
 openai==1.70.0